Initial Commit - Simple Read-Only API
This is a simple, but functional, proof of concept of a read-only WebSockets wrapper for DroneKit. A Tornado-based websockets server runs on the drone, the transmitter, or another device connected to the drone's network with MAVLink access. WebSockets clients are given an API key+shared secret combination that allows them access to the API. There is a simple example JS library (if you can even call it a library) that allows you to manipulate attributes that are continuously updated from the DroneKit server in a JSON-centric sort of way.